+++ src/ts/components/core/table/TableTbody.tsx
@@ -3,9 +3,9 @@ import { TableElementProps } from "props/table";
 import React from "react";
 
 const TableTbody = (props: TableElementProps) => {
-    const { setProps, children, ...others } = props;
+    const { setProps, tableProps, children, ...others } = props;
 
-    return <Table.Tbody {...others}>{children}</Table.Tbody>;
+    return <Table.Tbody {...others} {...tableProps}>{children}</Table.Tbody>;
 };
 
 export default TableTbody;